% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bias.tif}
\alias{bias.tif}
\title{Bias data}
\format{
A Geotiff file. Geographic CRS: WGS84 (+proj=longlat +datum=WGS84 +no_defs).
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2021.12.09.471691v1}
}
\description{
Raster data with bias referring to sampling rates (residual bias) based on the
distance from occurrences to roads, rivers, airports and cities.
}
\examples{
(f <- system.file("raster/bias.tif", package = "bivarmap"))
terra::rast(f)

}
