% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivarmap_colmatrix.R
\name{bivarmap_colmatrix}
\alias{bivarmap_colmatrix}
\title{Creates color matrix for bivariate map plots}
\usage{
bivarmap_colmatrix(
  nbreaks = 3,
  breakstyle = "quantile",
  upperleft = "#0096EB",
  upperright = "#820050",
  bottomleft = "#BEBEBE",
  bottomright = "#FFE60F",
  xlab = "x label",
  ylab = "y label",
  plotLeg = TRUE,
  saveLeg = FALSE
)
}
\arguments{
\item{nbreaks}{\verb{[numeric(1)=3]} \cr Number of breaks for each dimension
of the matrix.}

\item{breakstyle}{\verb{[character(1)="quantile"]} \cr Option to break each dimension
of the matrix in class intervals to set the colors. It can assume any of the following
values: "fixed", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust",
"fisher", "jenks", "dpih" or "headtails". (not sd???) For more information, see
\code{\link[classInt:classIntervals]{classInt::classIntervals()}}.}

\item{upperleft, upperright, bottomleft, bottomright}{\verb{[character(1)]} \cr
Colors to be assigned to the upperleft, upperright, bottomleft, and bottomright
corners of the color matrix. It can be a string with the name of a color or a
hexadecimal values representing the color.}

\item{xlab, ylab}{\verb{[character(1)]} \cr Label for the color matrix plot axes, x and y.}

\item{plotLeg}{\verb{[logical(1)=TRUE]} \cr If \code{TRUE}, a preview of the legend is plotted.}

\item{saveLeg}{\verb{[logical(1)=TRUE]} \cr Should the legend be saved in an external file?
Default is \code{FALSE}.}
}
\value{
Matrix of colors to be used in bivariate map plots.
}
\description{
This function creates the matrix of colors to be used in
bivariate maps plots. It created a matrix with \code{nbreaks} x \code{nbreaks}
colors do be used to represent a gradient in two different
continuous variables.
}
\examples{
data("temprec")

colmatrix <- bivarmap::bivarmap_colmatrix(nbreaks = 9)
colmatrix

}
