% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivarmap_map.R
\name{bivarmap_map}
\alias{bivarmap_map}
\title{Plots bivariate maps}
\usage{
bivarmap_map(
  bivarmap,
  colmatrix,
  x_legend_pos = 0,
  y_legend_pos = 0,
  x_legend_size = 0.2,
  y_legend_size = 0.2
)
}
\arguments{
\item{bivarmap}{\verb{[RasterLayer,SpatRaster]} \cr Bivariated classified
map, such as one generated by \code{\link[=bivarmap_raster]{bivarmap_raster()}}.}

\item{colmatrix}{\verb{[matrix]} \cr Matrix of colors to be used in the raster
classification and plot, created with \code{\link[=colmatrix]{colmatrix()}}.}
}
\value{
A \code{ggplot} plot of the bivariate map, with colors following a
bivariate color matrix, which is optionally also plotted.
}
\description{
This function plots bivariate maps with their color matrix
associated with gradients in the two variables or dimensions.
}
\examples{
# package
library(bivarmap)

# data
data("temprec")
temprec

# color matrix
colmatrix <- bivarmap::bivarmap_colmatrix(nbreaks = 9,
                                          xlab = "Temperatura",
                                          ylab = "Precipitação")
colmatrix

# raster
raster_col <- bivarmap::bivarmap_raster(rasterx = temprec$temp,
                                        rastery = temprec$prec,
                                        colmatrix = colmatrix)
raster_col

# plot map
bivarmap::bivarmap_map(bivarmap = raster_col,
                       colmat = colmatrix,
                       x_legend_pos = .2,
                       y_legend_pos = .1,
                       x_legend_size = .3,
                       y_legend_size = .3)
#--------------
# example sarbeco
library(terra)

rasts_folder <- system.file("raster", package = "bivarmap")

# open hosts raster
ho <- system.file("raster/hosts.tif", package = "bivarmap")
hosts <- terra::rast(ho)

# open bias raster
bi <- system.file("raster/bias.tif", package = "bivarmap")
bias <- terra::rast(bi)

# resample, since the maps have different resolution
b <- terra::resample(hosts, bias) \%>\%
    c(bias)

# set projection
terra::crs(b) <- '+proj=longlat +datum=WGS84 +no_defs '

# transorm data to proportions
b[[1]] <- b[[1]]/ max(na.omit(values(b[[1]])))
b[[2]] <- b[[2]]/ max(na.omit(values(b[[2]])))

# plot separate rasters
plot(b, col = viridis::viridis(10))

# color matrix
colmatrix <- bivarmap::bivarmap_colmatrix(nbreaks = 10,
                                          upperleft = "cyan",
                                          upperright = "purple",
                                          bottomleft = "beige",
                                          bottomright = "brown1",
                                          xlab = names(b)[1],
                                          ylab = names(b)[2])
colmatrix

# raster
raster_col <- bivarmap::bivarmap_raster(rasterx = b[[1]],
                                        rastery = b[[2]],
                                        colmatrix = colmatrix)
raster_col

# map
bivarmap::bivarmap_map(bivarmap = raster_col,
                       colmat = colmatrix,
                       x_legend_pos = .01,
                       y_legend_pos = .1,
                       x_legend_size = .2,
                       y_legend_size = .2)
}
