% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivarmap_raster.R
\name{bivarmap_raster}
\alias{bivarmap_raster}
\title{Creates a bivariate raster}
\usage{
bivarmap_raster(
  rasterx,
  rastery,
  colmatrix = col.matrix,
  export.colour.matrix = TRUE,
  outname = paste0("colMatrix_rasValues", names(rasterx))
)
}
\arguments{
\item{rasterx}{\verb{[RasterLayer,SpatRaster]} \cr RasterLayer representing the
x (first) axis or variable of the bivariate plot.
Support to \code{SpatRaster} to be added.}

\item{rastery}{\verb{[RasterLayer,SpatRaster]} \cr RasterLayer representing the
y (second) axis or variable of the bivariate plot.
Support to \code{SpatRaster} to be added.}

\item{colmatrix}{\verb{[matrix]} \cr Matrix of colors to be used in the raster
classification and plot, created with \code{\link[=colmatrix]{colmatrix()}}.}
}
\value{
A \code{RasterLayer} representing the bivariate map classified
accordingly with selected classes for each variable, for bivariate
map plotting.
}
\description{
Description... documentation to be continued.
}
\examples{
examples/bivarmap_raster_example.R

}
