% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hosts.tif}
\alias{hosts.tif}
\title{Hosts spatial data}
\format{
A Geotiff file. Geographic CRS: WGS84 (+proj=longlat +datum=WGS84 +no_defs).
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2021.12.09.471691v1}
}
\description{
Global raster data with Sarbecovirus bat host species number estimated
by ecological niche models for the present.
}
\examples{
(f <- system.file("raster/hosts.tif", package = "bivarmap"))
terra::rast(f)

}
