% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temprec}
\alias{temprec}
\title{Temperature and precipitation data}
\format{
An object of class \code{RasterBrick} of dimension 234 x 236 x 2.
}
\source{
\url{where_does_it_com_from.com}
}
\usage{
temprec
}
\description{
Describe here...
}
\examples{
data("temprec")
temprec

}
\keyword{datasets}
