% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivarmap_raster.R
\name{bivarmap_raster}
\alias{bivarmap_raster}
\title{Creates a bivariate raster}
\usage{
bivarmap_raster(
  rasterx,
  rastery,
  colmatrix = col.matrix,
  export.col.matrix = TRUE,
  outname = paste0("colMatrix_rasValues", names(rasterx))
)
}
\arguments{
\item{rasterx}{\verb{[RasterLayer,SpatRaster]} \cr \code{RasterLayer} object (from \link{raster} package)
or \code{SpatRaster} (from \link{terra} package) representing the x (first) axis or variable
of the bivariate plot.}

\item{rastery}{\verb{[RasterLayer,SpatRaster]} \cr \code{RasterLayer} object (from \link{raster} package)
or \code{SpatRaster} (from \link{terra} package) representing the y (second) axis or variable
of the bivariate plot.}

\item{colmatrix}{\verb{[matrix]} \cr Matrix of colors to be used in the raster
classification and plot, created with \code{\link[=colmatrix]{colmatrix()}}.}

\item{export.col.matrix}{To be documented.}

\item{outname}{To be documented}
}
\value{
A \code{RasterLayer} representing the bivariate map classified
accordingly with selected classes for each variable, for bivariate
map plotting.
}
\description{
Description... documentation to be continued.
}
\examples{
examples/bivarmap_raster_example.R

}
