% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hosts.tif}
\alias{hosts.tif}
\title{Hosts spatial data}
\format{
A Geotiff file. Geographic CRS: WGS84 (+proj=longlat +datum=WGS84 +no_defs).
\itemize{
\item{1:} {Presence of cabins}
\item{NA:} {No presence of cabins}
}
}
\source{
\url{to_complete_here}
}
\description{
Global raster data with info from hosts...
}
\examples{
(f <- system.file("raster/hosts.tif", package = "bivarmap"))
raster::raster(f)

}
